Moves tests for get_year_links() to separate file
